+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 WooCommerce for golang Change Log
 =================================
 
-## 1.0.2 under development
+## 1.0.2
 
 - Bug: Fixed parse string to float64 failed if an empty string
 - New: Added data and report service tests